When backtracking, zero the region(s) for which all possible combinations have been tried.
Zeroing the regions restores the board to the state where no combinations
have (yet) been tried for the region in question.  If regions are not
restored it seems plausible that the validity of row/column being checked
could be affected by the residue from previous attempts at finding a solution.

Also added a couple of "hard" sudoku puzzle Google search turned up for me as test cases.